Pprobsssss

+++ Order.java
@@ -60,7 +60,7 @@ public class Order extends PcParts{
 	public long getPhone(){
 		return phone;
 	}	
-	public void setPhone(int phone){
+	public void setPhone(long phone){
 		this.phone = phone;
 	}
